% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_model.R
\name{BC_model}
\alias{BC_model}
\title{Predict Rank Abundance Distribution}
\usage{
BC_model(df_abundance = NULL, column = NULL, BC_param_object = NULL)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{BC_param_object}{Optional. A previous object generated by \code{BC_param}.}
}
\value{
A list with the following elements: The input data frame with added processed ranking data, model data and confidence interval data, the adjusted parameters, the confidence interval of the parameters, the linear model, a summary of the model and a generated function for use with raw numeric data.
}
\description{
Uses the parameters calculated by \code{BC_param} to model a DGBD distribution from the rank information in the data frame.
}
\examples{
BC_model(datasets::iris,"Petal.Length")

BC_model(datasets::PlantGrowth,1)

BC_model(datasets::ToothGrowth,"len")
}
