% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_param.R
\name{BC_param}
\alias{BC_param}
\title{Estimate the DGBD Distribution Parameters}
\usage{
BC_param(df_abundance, column, confidence_interval = 0.95)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{confidence_interval}{The confidence interval to calculate for the DGBD distribution.}
}
\value{
A list with the following elements: The input data frame with added processed ranking data and predicted confidence interval values, the adjusted parameters, the confidence interval of the parameters, the linear model and a summary of the model.
}
\description{
\code{BC_param} calculates the log of the data and estimates the abundance from the ranking using a linear model.
The coefficients of the linear model are then scaled for posterior use.
}
\examples{
BC_param(datasets::iris,"Petal.Length")

BC_param(datasets::PlantGrowth,1,confidence_interval=0.99)

BC_param(datasets::ToothGrowth,"len",confidence_interval=0.999)
}
