% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_report.R
\name{BC_report}
\alias{BC_report}
\title{Summarize the Modeled DGBD Distribution}
\usage{
BC_report(
  df_abundance = NULL,
  column = NULL,
  BC_plot_object = NULL,
  show_plot = T,
  show_stats = T
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{BC_plot_object}{A previous object generated by \code{BC_plot}.}

\item{show_plot}{Logical. Whether to plot the DGBD distribution. Defaults to true.}

\item{show_stats}{Logical. Whether to print the parameters of the DGBD distribution and its R2 to the console. Defaults to true.}
}
\value{
A list with the following elements: a matrix that contains the parameters used and their confidence intervals, the R2 and the DGBD plot.
}
\description{
Summarize the Modeled DGBD Distribution
}
\examples{
BC_report(datasets::iris,1)

BC_report(datasets::PlantGrowth,"weight",show_plot=FALSE)

BC_report(datasets::ToothGrowth,"len",show_stats=FALSE)
}
