% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radbc-package.R
\docType{package}
\name{radbc-package}
\alias{radbc}
\alias{radbc-package}
\title{radbc: Proof of Concept R Wrapper for 'ADBC'}
\description{
Provides a proof-of-concept R wrapper for 'Arrow' Database Connectivity ('ACBC').
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
