% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radbc.R
\name{radbc_connection_init}
\alias{radbc_connection_init}
\alias{radbc_connection_init_default}
\alias{radbc_connection_set_options}
\alias{radbc_connection_release}
\title{Connections}
\usage{
radbc_connection_init(database, ...)

radbc_connection_init_default(database, options = NULL, subclass = character())

radbc_connection_set_options(connection, options)

radbc_connection_release(connection)
}
\arguments{
\item{database}{An \link[=radbc_database_init]{radbc_database}.}

\item{...}{Driver-specific options. For the default method, these are
named values that are converted to strings.}

\item{options}{A named \code{character()} or \code{list()} whose values are converted
to strings.}

\item{subclass}{An extended class for an object so that drivers can specify
finer-grained control over behaviour at the R level.}

\item{connection}{An \link[=radbc_connection_init]{radbc_connection}}
}
\value{
An object of class 'radbc_connection'
}
\description{
Connections
}
\examples{
db <- radbc_database_init(radbc_driver_void())
radbc_connection_init(db)

}
