% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radbc.R
\name{radbc_database_init}
\alias{radbc_database_init}
\alias{radbc_database_init_default}
\alias{radbc_database_set_options}
\alias{radbc_database_release}
\title{Create an ADBC Database}
\usage{
radbc_database_init(driver, ...)

radbc_database_init_default(driver, options)

radbc_database_set_options(database, options)

radbc_database_release(database)
}
\arguments{
\item{driver}{An radbc_driver}

\item{...}{Driver-specific options. For the default method, these are
named values that are converted to strings.}

\item{options}{A named \code{character()} or \code{list()} whose values are converted
to strings.}
}
\value{
An object of class radbc_database
}
\description{
Create an ADBC Database
}
\examples{
radbc_database_init(radbc_driver_void())

}
