% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_broken_record.R
\name{radbc_driver_broken_record}
\alias{radbc_driver_broken_record}
\title{Always return the same data}
\usage{
radbc_driver_broken_record(data)
}
\value{
An object of class 'radbc_driver_broken_record'
}
\description{
A variant of the broken_record driver that always returns the same data set.
}
\examples{
db <- radbc_database_init(radbc_driver_broken_record(mtcars))
con <- radbc_connection_init(db)
stmt <- radbc_statement_init(con)
result <- radbc_statement_execute_query(stmt)
as.data.frame(result$get_next())

}
