% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_monkey.R
\name{radbc_driver_monkey}
\alias{radbc_driver_monkey}
\title{Monkey see, monkey do!}
\usage{
radbc_driver_monkey()
}
\value{
An object of class 'radbc_driver_monkey'
}
\description{
A driver whose query results are set in advance.
}
\examples{
db <- radbc_database_init(radbc_driver_monkey())
con <- radbc_connection_init(db)
stmt <- radbc_statement_init(con, mtcars)
result <- radbc_statement_execute_query(stmt)
as.data.frame(result$get_next())

}
