% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_void.R
\name{radbc_driver_void}
\alias{radbc_driver_void}
\alias{radbc_driver}
\title{Create ADBC drivers}
\usage{
radbc_driver_void()

radbc_driver(driver_init_func, ..., subclass = character())
}
\arguments{
\item{driver_init_func}{An external pointer to a DL_FUNC with the type
\code{AdbcDriverInitFunc} specified in the adbc.h header. The external pointer
must have the class "radbc_driver_init_func".}

\item{...}{Further key/value parameters to store with the (R-level) driver
object.}

\item{subclass}{An optional subclass for finer-grained control of
behaviour at the R level.}
}
\value{
An object of class 'radbc_driver'
}
\description{
Creates the R object representation of an ADBC driver, which consists of a
name and an initializer function with an optional subclass to control
finer-grained behaviour at the R level.
}
\examples{
radbc_driver_void()

}
