% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radbc.R
\name{radbc_statement_init}
\alias{radbc_statement_init}
\alias{radbc_statement_init_default}
\alias{radbc_statement_set_options}
\alias{radbc_statement_release}
\title{Statements}
\usage{
radbc_statement_init(connection, ...)

radbc_statement_init_default(
  connection,
  options = NULL,
  subclass = character()
)

radbc_statement_set_options(statement, options)

radbc_statement_release(statement)
}
\arguments{
\item{connection}{An \link[=radbc_connection_init]{radbc_connection}}

\item{...}{Driver-specific options. For the default method, these are
named values that are converted to strings.}

\item{options}{A named \code{character()} or \code{list()} whose values are converted
to strings.}

\item{subclass}{An extended class for an object so that drivers can specify
finer-grained control over behaviour at the R level.}

\item{statement}{An \link[=radbc_statement_init]{radbc_statement}}
}
\value{
An object of class 'radbc_statement'
}
\description{
Statements
}
\examples{
db <- radbc_database_init(radbc_driver_void())
con <- radbc_connection_init(db)
radbc_statement_init(con)

}
