% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radbc.R
\name{radbc_statement_set_sql_query}
\alias{radbc_statement_set_sql_query}
\alias{radbc_statement_set_substrait_plan}
\alias{radbc_statement_prepare}
\alias{radbc_statement_get_parameter_schema}
\alias{radbc_statement_bind}
\alias{radbc_statement_bind_stream}
\alias{radbc_statement_execute_query}
\title{Statement methods}
\usage{
radbc_statement_set_sql_query(statement, query)

radbc_statement_set_substrait_plan(statement, plan)

radbc_statement_prepare(statement)

radbc_statement_get_parameter_schema(statement)

radbc_statement_bind(statement, values, schema = NULL)

radbc_statement_bind_stream(statement, stream, schema = NULL)

radbc_statement_execute_query(statement)
}
\arguments{
\item{statement}{An \link[=radbc_statement_init]{radbc_statement}}

\item{query}{An SQL query as a string}

\item{plan}{A raw vector representation of a serialized Substrait plan.}

\item{values}{A \link[nanoarrow:as_nanoarrow_array]{nanoarrow_array} or object
that can be coerced to one.}

\item{schema}{A \link[nanoarrow:as_nanoarrow_schema]{nanoarrow_schema} or object
that can be coerced to one.}

\item{stream}{A \link[nanoarrow:as_nanoarrow_array_stream]{nanoarrow_array_stream}
or object that can be coerced to one.}
}
\value{
\itemize{
\item \code{radbc_statement_set_sql_query()}, \code{radbc_statement_set_substrait_plan()},
\code{radbc_statement_prepare()}, \code{radbc_statement_bind()},
\code{radbc_statement_bind_stream()}, and \code{radbc_statement_execute_query()}
return \code{statement}, invisibly.
\item \code{radbc_statement_get_parameter_schema()} returns a
\link[nanoarrow:as_nanoarrow_schema]{nanoarrow_schema}.
}
}
\description{
Statement methods
}
\examples{
db <- radbc_database_init(radbc_driver_void())
con <- radbc_connection_init(db)
stmt <- radbc_statement_init(con)
# (not implemented by the void driver)
try(radbc_statement_set_sql_query(stmt, "some query"))

}
